+++ client/package.json
@@ -10,9 +10,7 @@
     "axios": "^0.21.4",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-redux": "^8.0.2",
     "react-scripts": "4.0.3",
-    "redux": "^4.2.0",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
