--- client/src/lib/store.js
@@ -1,6 +1,0 @@
-import { createStore } from "redux";
-import rootReducer from "./reducers/rootReducer";
-
-const store = createStore(rootReducer)
-
-export default store
